Rename typing.py to pgtypes.py to make the import more readable.
