+++ pygsti/__init__.py
@@ -7,7 +7,7 @@
 # http://www.apache.org/licenses/LICENSE-2.0 or in the LICENSE file in the root pyGSTi directory.
 #***************************************************************************************************
 """ A Python implementation of LinearOperator Set Tomography """
-from .typing import (
+from .pgtypes import (
     SpaceT
 )
