+++ pygsti/pgtypes.py
